Delegator refactoring: Endgame request a piece from multiple peers.
git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@182 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/content/delegator_piece.h
@@ -2,6 +2,8 @@
 #define LIBTORRENT_DELEGATOR_PIECE_H
 
 #include <vector>
+#include <inttypes.h>
+
 #include "piece.h"
 
 namespace torrent {
@@ -31,9 +33,9 @@ public:
   const Piece&       get_piece()                        { return m_piece; }
   void               set_piece(const Piece& p)          { m_piece = p; }
 
-  unsigned int       get_reservees_size()               { return m_reservees.size(); }
-  unsigned int       get_stalled()                      { return m_stalled; }
-  unsigned int       get_not_stalled()                  { return m_reservees.size() - m_stalled; }
+  uint32_t           get_reservees_size()               { return m_reservees.size(); }
+  uint16_t           get_stalled()                      { return m_stalled; }
+  uint16_t           get_not_stalled()                  { return m_reservees.size() - m_stalled; }
 
 protected:
   friend class DelegatorReservee;
@@ -51,7 +53,7 @@ private:
   Reservees          m_reservees;
 
   bool               m_finished;
-  unsigned int       m_stalled;
+  uint16_t           m_stalled;
 };
 
 }
